+++ src/render/minimap.c
@@ -6,7 +6,7 @@
 /*   By: jsong <jsong@student.codam.nl>               +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2025/05/09 15:21:43 by jsong         #+#    #+#                 */
-/*   Updated: 2025/05/19 17:53:41 by jianisong     ########   odam.nl         */
+/*   Updated: 2025/05/21 15:23:05 by jsong         ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,8 +27,8 @@ static void	draw_minimap_cell(t_game *game, int i, int j)
 	int		my;
 	t_cell	cell;
 
-	mx = game->player.x - MINIMAP_RADIUS + i;
-	my = game->player.y - MINIMAP_RADIUS + j;
+	mx = (int)game->player.x - MINIMAP_RADIUS + i;
+	my = (int)game->player.y - MINIMAP_RADIUS + j;
 	if (is_valid_point(&game->map, mx, my))
 	{
 		cell.px = i * MINIMAP_CELL_SIZE;
